pkg-export-k8s: Generate v1beta2 custom version of habitat manifest
This only changes the version of the manifest the exporter will
generate, but does not add support for generating the new fields in
the manifest yet. The only difference with the previous version of the
generated manifest is that the Habitat operator will use Kubernetes
stateful sets instead of deployments as a way to deploy services in
the cluster.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>