pkg-export-k8s: Add a quote helper to manifest generator
We will want to generate fields with arbitrary values in the YAML
document, so we need to put them inside quotes. For this to work, the
string inside the quotes needs to have backslashes and quotes
escaped. The new helper does exactly that - escapes the necessary
chars and wraps the string in quotes.

This will be needed for generating environment variables fields in
Kubernetes manifest.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>