Minimal change to make `make lint` successfully run clippy checks
Currently all failing lints in the default set are allowed to give a clean
baseline. As we work through the lints to discover which are useful, they
will be changed to causing errors and eventually made part of the automated
tests if that proves useful.

Since this changes the method for running clippy from an optional dependency
to a cargo subcommand, we can subsequently remove the dependency code.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>