Switch to `prost` from `protobuf` crate in Butterfly
This is a sister change to #5126 which removed the protobuf crate and
replaced it with the prost crate in the Supervisor proper. These changes
are house cleaning work in preparation of improving Butterfly and the
Supervisor overall.

In addition to the refactor, this change will fix a number of
serialization/deserialization bugs which were introduced at some point
into Butterfly's types.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>