Add caching of $HOME/pkgs back
This should speed up the build significantly as this is where the builds for
libarchive, libsodium and zmq live. They should be independent of the rust
toolchain, so hopefully there will be no issues caching them across nightly
and the stable versions used for other travis jobs

It looks like the issue was that LD_LIBRARY_PATH lacked the reference to the
nightly toolchain in the case that the .cargo dir was cached. Perhaps this is
updated as a part of the regular install (or the clippy install). It may be
worth checking whether adding --force to `cargo install clippy` would help.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>