Add more debugging prints around librustc_driver*.so issue
Even though the appropriate ~/.multirust/toolchains/nightly-*/lib was added to
LD_LIBRARY_PATH, the error recurred when caching was turned on. It appears that
there is a version mismatch or something. The error says:

/home/travis/.cargo/bin/clippy-driver: error while loading shared libraries: librustc_driver-1fbd3f11ce4eb66a.so: cannot open shared object file: No such file or directory

But the find output shows:

/home/travis/.multirust/toolchains/nightly-x86_64-unknown-linux-gnu/lib/librustc_driver-08fcc8f32922954d.so

What is the explanation for 1fbd3f11ce4eb66a vs 08fcc8f32922954d?

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>