Try to confirm that the error is related to caching the .cargo directory
The error occurs when running .cargo/bin/clippy-driver. It's looking for a
different version of librustc_driver-*.so than is installed. What's confusing
is that when run locally, `cargo clippy` doesn't even touch clippy-driver, so
I'm not sure what's going on differently here, or exactly how clippy interacts
with rustc et al.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>