pkg-export-*: Use failure-related crates from crates registry
In the beginning I specified the git field to point to the original
author and the version constraint to "~1" in Cargo.toml, but then I
found out that some prost package uses failure from crates registry,
so I opted for downgrading to just use a single version of the crates.

On top of that we were pulling failure crates from the git repo of the
original author. But now failure is a part of rust lang nursery. I
expect that the code from nursery gets packaged and pushed into crates
registry. Admittedly for some reason, the newest version available
from git wasn't yet pushed to the registry.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>