Fix serialization problem where defaults were injected
Prost will incorrectly concat the default value specified in a protocol
file to the value of a field if the field is marked as required. This
change removes the default values that were placed in the protocol file
for `types.proto` to side step the issue, but we should follow up with
the creators of Prost to get this solved for real. We weren't using the
default value in the protocol file anyway, so this isn't much of a
change.

Signed-off-by: Jamie Winsor <jamie@vialstudios.com>