Make querying an unloaded service status always exit nonzero
Resolves https://github.com/habitat-sh/habitat/issues/5172

Previously, an early exit in the case of no loaded services caused a successful
exit even if there was an <IDENT> argument. This undermined the value of
`hab svc status <IDENT>` as a check on a service being loaded.

Also, this eliminates one cause of the crash avoided by
https://github.com/habitat-sh/habitat/pull/5166.
Calling CtlRequest::reply_complete multiple times resulted in an inconsistent
state for the request. That fix should remain, since it's preferable to not
crash, but it illustrates why the CtlRequest.tx debug output was confusing.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>