Make launcher exit status indicate if supervisor exited with a fatal error
If the supervisor exits ERR_NO_RETRY_EXCODE (86), the launcher's exit status
determines the exit status of `hab sup run`, which should be nonzero to
communicate that the supervisor didn't start successfully.

Also, make the supervisor process exit ERR_NO_RETRY_EXCODE in the event of
ButterflyError. This includes a failure to bind to the supervisor port on
startup.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>