Pin everything to use pre-refresh core plans
The refresh brought in a version of glibc that dropped support for
older 2.6 Linux kernels. That is currently being sorted out, but until
then, this pins all core dependencies to the pre-refresh versions to
ensure that our next release continues to run on older kernels.

While some dependencies could have been pinned at the version level, I
felt it prudent to be a pinning maximalist in this case. As a result,
everything except dependencies on actual Habitat packages is pinned
with a fully-qualified identifier. (If we pinned those, we wouldn't be
able to make a release!)

Signed-off-by: Christopher Maier <cmaier@chef.io>