Loosen up pins on Docker exporter.
Initially, we locked everything in this repository back to pre-"base
plans refresh" dependencies when it was discovered that the new
`glibc` dropped support for older kernels.

In hindsight, that was more conservative than we needed to be. In
particular, locking down the Docker exporter ended up unnecessarily
breaking our Builder development environment. The exporter is
statically compiled, and in any case, Docker doesn't run on older
kernels anyway.

Additional packages can be relaxed in future PRs, but this will
unblock additional work in Builder right now.

Signed-off-by: Christopher Maier <cmaier@chef.io>