Relax a few pins
We can relax some pins for things that are statically compiled, as
well as for our build dependencies.

The whole reason to pin was to continue to allow things to run on
older kernels following the June 2018 base plans refresh, but
statically-compiled binaries and buildtime dependencies don't apply
here.

It does mean that these things won't be able to *build* on older
kernels, but we don't build on older kernels.

The Kubernetes and Helm packages are a little different, in that
they *do* have runtime dependencies, but these dependencies (docker
and helm) do not have dependencies themselves. Also, we still require
a Docker daemon to run these exporters, and those don't run on older
kernels anyway.

Signed-off-by: Christopher Maier <cmaier@chef.io>