[core] Reimplement PackageTarget as a type-driven unit type.
This change updates the internal implementation of the `PackageTarget`
type while preserving all previous useful external behavior. Note that
this rework was not designed to be a 100% drop-in replacement, but
rather a revisiting of `PackageTarget`'s public API.

The main goal of this work was to enforce a strongly atomic, indivisible
concept of a package target that didn't leak its internal representation
more than necessary. This approach uses Rust's type system (driven
internally with an `enum`) to enforce safety and determinism while still
exposing one canonical string representation for each target type.

The majority of the developer documentation for the `package::target`
module and the `PackageType` struct is in rustdoc format, so it's highly
recommended to run `cargo doc --open` in the `components/core` directory
to see the fully formatted docs and examples.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>