[core] Refactor `PackageIdent#archive_name()` & update return type.
This change refactors the internals of the `#archive_name()` and
`#archive_name_with_target()` methods on `PackageIdent` to use the
`Display` trait implementations of `PackageIdent` and `PackageTarget`.

Additionally the return type on these functions were updated from
`Option<String>` to `Result<String` as returning an option hides the
underlying source of the error which leads to the current return of
`None`, namely that the current package identifier is *not* fully
qualified but should be.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>