[core] Only load same active target with `PackageInstall`.
This change affects the implementation of the `load()` and
`load_at_least()` functions for `PackageInstall`. Prior to this change,
an installed package would be determined given a `PackageIdent`, but the
check for whether the resulting installed package's `PackageTarget`
matched the builtin target happened after the "correct" candidate was
determined. Consequently, a failure would be returned and leave a user
with no recourse.

In order to prevent such failure scenarios, the error-preventing
validation was moved into the logic which generates the list of
candidate package, called from `package_list()`. Specifically, the
`walk_releases()` function iterates through a collection of paths that
should correspond to package prefixes. Rather than return all matching
paths as "installed packages", this change adds a crucial check to make
sure that the candidate's package target matches the code's active
target. Unless this check is successful, the candidate is skipped and
thus filtered out of consideration.

As a result, there are new conditions where a directory that appears to
be an installed package might not be considered as a match for a
`PackageIdent`:

* If the installed package doesn't have a `TARGET` metafile in the
package's root directory (or if the `TARGET` file cannot be read)
* If the contents of the `TARGET` metafile couldn't be parsed into a
supported `PackageTarget`
* If the installed package's parsed `PackageTarget` doesn't match the
active target, checked with `PackageTarget::active_target()`.

If any of the above conditions are met, a debugging message is generated
(available if the user has the `RUST_LOG` environment variable set), and
the candidate is skipped over.

In order to ensure the correct behavior is captured and preserved,
several units tests were added to help better document these scenarios.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>