Update relevant Habitat Plans to set PLAN_PACKAGE_TARGET.
This change explicitly injects a package target value for Rust software
components that deal with Habitat packages. The `PLAN_PACKAGE_TARGET`
environment variable is consumed *at build time* by an
`active_package_target` function in the `habitat_core::package::target`
module (which is used once to set a static value at runtime). This is
the mechanism to "bake in" a package target for our tooling.

It should be noted that the `PLAN_PACKAGE_TARGET` is **not** a public,
Habitat environment variable. It is an internal Habitat source
code-consuming variable that is not referenced or used anywhere else in
our system (by design).

Additionally, this approach of setting `PLAN_PACKAGE_TARGET` will not be
used for Builder software components **unless** we plan to build Builder
packages for platform targets other than `x86_64-linux`.

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>