Fix protocol build process
When building a Habitat package, we are not able to use the `protoc`
binary that is included in `prost-build`, because it is dynamically
linked, making it fail in our studio-based builds. We'll use
`core/protobuf` as the source of our `protoc` binary instead.

Similarly, we cannot simply rename the generated Rust source code
because the source and destination are on different filesystems within
a studio. Instead, we need to copy the files to their final destination.

We *could* simply include the generated files directly in the code
that uses them, but for now, we will continue to check them into version
control, and include _those_ files.

That is to say, instead of doing something like:

``` rust
include!(concat!(env!("OUT_DIR"), "/butterfly.newscast.rs"));
```

we are opting to do this instead:

``` rust
include!("../generated/butterfly.newscast.rs");
```

The former method would not include the generated code in version
control. We are explicitly choosing to commit the generated code in
order to make it easier to spot changes during code review, and to
generally reason about what is going on.

(This is what the code did before this commit, but it's worth calling
it out explicitly.)

Signed-off-by: Christopher Maier <cmaier@chef.io>