Avoid incomplete package installs
Previously, `hab pkg install` would unpack a package directly into the
target directory. If `hab pkg install` was killed in the middle of
the installation, the system might be left in state where `hab` would
report the package as installed, even though it was unusable.

This changes the unpack process so that we first unpack into a
temporary directory and then rename the directory into place.

The trade-off here is that a killed install will leave behind a
temporary directory that would take up space on the user's disk.

Signed-off-by: Steven Danna <steve@chef.io>