Remove release build triggering from TravisCI
While we no longer need to build releases in Travis, we can still use
it to build packages following PR merges to master.

We've simplified the the old Mac release scripts, too. We only ever
push to the unstable repo with these scripts, since we're no longer
building an official release here in TravisCI. Therefore, we don't
need logic to determine which repo to push to.

Signed-off-by: Christopher Maier <cmaier@chef.io>