If the supervisor terminates with an error result, don't restart it
Resolves https://github.com/habitat-sh/habitat/issues/3060

Exit ERR_NO_RETRY_EXCODE to communicate to the launcher that it shouldn't be
restarted. Otherwise errors on startup can lead to a tight loop of perpetual
restarts.

Abrupt exits due to panics and graceful supervisor exit for update will still
exit 0 and be restarted by the launcher.

Also, add a test script.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>