Create a SelfUpdater, even if the supervisor version isn't fully qualified
The only case when the supervisor version won't be fully qualified
is when testing a cargo-built binary and running it with the HAB_SUP_BINARY
override mechanism. Removing the fully qualified check allows for easier
testing of the update mechanism with the cargo-built binary. There should be
no impact on officially released builds.

In the case of cargo-built binaries, since the version lacks the release
number, it will always appear to be older than releases which contain it. This
allows the update logic to be triggered, but upon restart, the HAB_SUP_BINARY
override is still in effect, so the supervisor will not change.

This also simplifies the logic since self_updater no longer needs to be wrapped
in an option, so it's a generally beneficial change.

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>