Promote packages to a separate channel before promoting to stable
The idea is that Builder will consume from `builder-live`, and once
we're satisfied there, then we promote the packages to `stable` for
consumption by the wider world.

This is intended as a temporary measure only; our acceptance
environment needs to be beefed up in order to provide a better
early-warning for breaking changes. That work is in progress, but
until it's here, this provides a little more safety for us.

Signed-off-by: Christopher Maier <cmaier@chef.io>