Allow Docker studio image name to be overridden
Previously, the image was always created based on our Bintray
repository. This makes it more complicated than it should be to, say,
generate development images and push them to a separate
repository (yes, yes, you could always retag if you really wanted to).

This change allows us to set `IMAGE_NAME` in the environment if we
wish to generate the image with a different image name. If that
variable is not set, we'll continue to use the standard Bintray
repository.

Before:
``` sh
$ IMAGE_NAME=chuckleheads/studio HAB_BLDR_CHANNEL=stable ./build-docker-image.sh
...
Successfully tagged habitat-docker-registry.bintray.io/studio:0.59.0-20180712161521
--    build-docker-image.sh: Tagging latest image to habitat-docker-registry.bintray.io/studio:0.59.0-20180712161521
--    build-docker-image.sh: Tagging latest image to habitat-docker-registry.bintray.io/studio:0.59.0
--    build-docker-image.sh:
--    build-docker-image.sh: Docker Image: habitat-docker-registry.bintray.io/studio:0.59.0-20180712161521
--    build-docker-image.sh: Build Report: /home/maier/src/habitat-sh/habitat/components/studio/results/last_image.env
--    build-docker-image.sh:
```

After:
``` sh
$ IMAGE_NAME=chuckleheads/studio HAB_BLDR_CHANNEL=stable ./build-docker-image.sh
...
Successfully tagged chuckleheads/studio:0.59.0-20180712161521
--    build-docker-image.sh: Tagging latest image to chuckleheads/studio:0.59.0-20180712161521
--    build-docker-image.sh: Tagging latest image to chuckleheads/studio:0.59.0
--    build-docker-image.sh:
--    build-docker-image.sh: Docker Image: chuckleheads/studio:0.59.0-20180712161521
--    build-docker-image.sh: Build Report: /home/maier/src/habitat-sh/habitat/components/studio/results/last_image.env
--    build-docker-image.sh:
```

Signed-off-by: Christopher Maier <cmaier@chef.io>