Push "fake release" studio images to a separate repository
Previously, we were pushing everything into the same repository; the
only thing saving us was that we incorporate the version into the
image when we pull, which provides the needed isolation.

Now, we specifically override the image name to point to the
`studio-dev` repository when doing a fake release;

We also drop the `-r` option, since it's actually a no-op here.

Signed-off-by: Christopher Maier <cmaier@chef.io>