Don't quote HAB_STUDIO_SUP
If the variable isn't set, we end up passing "" as an argument, which
is interpreted by the Supervisor as a package identifier, which is not
legal.

This results in the Supervisor not being able to start properly when
entering a Studio. Manually running `sup-run` afterward starts things
up with no issues, though.

Signed-off-by: Christopher Maier <cmaier@chef.io>