[butterfly] Make DatFile write more durable
This adds two calls to fsync in the DatFile write method:

1) On the temporary file before the rename
2) On the parent directory after the rename

The first fsync is to persist the content of the dat file to disk. I
think this might have prevented of the 0-byte DatFile seen in #5460.

The second fsync is to persist the rename operation to disk. I've
no-oped it on Windows since we found that calling sync_all on a
directory in windows produces an error in
b688d8ae25d9f38b4053fd29372e2733c9254327.

Signed-off-by: Steven Danna <steve@chef.io>