Don't segfault `plan init` on io::Error
Fixes #5194.

Short story - we have some code that checks for what kind of project you
are building, and attempts to auto-detect the right scaffold for you.
Hiding in there is a check for any file that has a `.go` extension; and
it's signature is `io::Result<bool>`.

The calling method expects to return just a bool - but rather than
handle the `io::Error`, we called an `expect` on it. Since the right
behavior would be to simply return `false` if we can't read the file
(because, clearly, we couldn't detect what kind of file it is), this
patch simply changes from `expect` to `or_else(false)`.

In the original bug report, it is noted that no backtrace is produced.
For future humans, you can find things like this with gdb pretty easily
if you know round-about where the source is:

```shell
$ rust-gdb --args /home/adam/src/habitat/target/debug/hab plan init foo
[gdb] break scaffolding.rs:83
[gdb] r
.. step through the code ..
```

Of course, you could also just grep for the expect string. But where is
the fun in that?

Love,
Adam

Signed-off-by: Adam Jacob <adam@chef.io>