Improve error message when supervisor unavailable
This fixes #5367.

Previously, when doing a `hab svc (load|unload)`, or any command that
requires remote control of the supervisor, if there was a network issue,
we would simply return the upstream description of the problem (most
often, something like `Connection refused (os error 111)`.

This wasn't a very user friendly error message. Now, on any `SvcClient`
command that returns an `SvcClientError::Io` error, we will print the
following:

```
✗✗✗
✗✗✗ Unable to contact the Supervisor.
✗✗✗
✗✗✗ If the Supervisor you are contacting is local, this probably means it is not running. You can run a Supervisor in the foreground with:
✗✗✗
✗✗✗ hab sup run
✗✗✗
✗✗✗ Or try restarting the Supervisor through your oerating systems init process.
✗✗✗
✗✗✗ Original error is:
✗✗✗
✗✗✗ Connection refused (os error 111)
✗✗✗
```

Which is, I think, much more helpful.

Signed-off-by: Adam Jacob <adam@chef.io>