Add a supervisor version requirement checked by the launcher
The launcher needs to make assumptions about the supervisor binary's behavior,
so now it can be tied to a semantic version test that is part of the launcher
binary. In case this code is broken, an escape valve is provided in the form of
the HAB_LAUNCH_NO_SUP_VERSION_CHECK environment variable.

Resolves https://github.com/habitat-sh/habitat/issues/5388

Signed-off-by: Jon Bauman <5906042+baumanj@users.noreply.github.com>