Streamline member insert logic, using ordered Health type
This refactoring takes advantage of the `Ord` and `PartialOrd`
implementations of `Health` to simplify the logic of when we choose to
accept and propagate a new Membership rumor.

All health updating now goes through `insert_health_by_id` (pending a
future rename); this allows us to consistently ensure that we are
tracking timestamps when a member is marked as `Suspect` or
`Confirmed` (since these can "time out" to `Confirmed` or `Departed`,
respectively). Previously, this functionality was unevenly distributed
through the code. In particular, the current code fixes a few
inconsistencies that were discovered:

1) If a rumor was completely new, we weren't tracking any timestamps
if it was `Suspect` or `Confirmed`.

2) If the incoming rumor had a higher incarnation, we didn't track any
timestamps. I believe the underlying assumption here was that if
you've got a higher incarnation, it's a rumor from the member in
question, and they'll be `Alive` (otherwise they couldn't increment
their own incarnation), and we don't track timestamps for `Alive`
members. While this is probably true, it doesn't always have to be. In
particular, we may be receiving a rumor from another member that the
new-incarnation is `Suspect` before we receive the initial `Alive` rumor
from the member itself. I don't know how likely this scenario is, but
nothing rules it out 100%.

3) If we propagated an `Alive`->`Confirmed` or a `Suspect`->`Confirmed` rumor,
we didn't record any timestamps to time out to `Departed`.

Additional refactorings within the `MemberList` module include
renaming variables, struct members, functions, etc. to more clearly
reflect intent. Functionality that existed in the `expire` thread is
consolidated in `MemberList` now, resulting in fewer overall functions
and less data sharing.

A whole bunch of unit tests were added, including tests around timing
out `Confirmed` membership rumors to `Departed`, which didn't exist
before. Insertion of membership rumor data, as well as transitions
between all possible combinations of `Health` are exhaustively tested
now, as well.

Signed-off-by: Christopher Maier <cmaier@chef.io>