Establish baseline tests for `Ord` and `PartialOrd` for `Health`
Membership health exists on a continuum of

    Alive -> Suspect -> Confirmed -> Departed

It turns out that the latest version of `prost` automatically derives
`Ord` and `PartialOrd` for Enums (thanks, `prost`!), so this commit
merely introduces tests that ensure they work like we expect.

We have some logic where we propagate rumors only if the rumored
health is strictly "greater than" (alternatively, "worse than") the
health we are already aware of. Taking advantatage of `Health`
ordering will allow us to refactor this logic in a more concise
manner in a future commit.

Signed-off-by: Christopher Maier <cmaier@chef.io>