Don't send out a Departed Membership rumor on Supervisor updates
We were sending out a Departed Membership rumor whenever the
Supervisor process was shut down. This is fine when we're actually
terminating the entire Supervisor and all its processes, because it
allows the rest of the network to effectively bypass the entire
suspicion process and mark the Supervisor as "gone".

If the Supervisor is going down because it's updating itself, however,
this causes unnecessary work and creates instability, because the
Supervisor is really coming right back. Due to an existing bug that
conflates Service health with Supervisor reachability, this behavior
exacerbates the instability, because other Services will view the
Services running on the updating Supervisor as utterly gone, even
though they continue running just fine during the update process.

This does not fix that conflation, but it will provide an increase in
stability and correctness.

Signed-off-by: Christopher Maier <cmaier@chef.io>