[api-client] Always set package target when making package API calls.
This change starts in the Builder API client and flows out from there
with the goal of always providing an appropriate package target when
calling package-related API endpoints. This means for these calls
(namely, `fetch_package` and `show_package`) a non-optional
`PackageTarget` is required. Depending on the calling code, this value
could be provided upstream (i.e. user-provided value, reading from a
config file, etc.) or the compiled program's
`PackageTarget::active_target()` should be used. In all cases, the
derived value will always be clear.

As mentioned before, this work started in `builder-api-client`, and
rippled up (in a good way) into the common install logic and then
minimally into the `hab` CLI, Supervisor call sites. Except for one
remaining location (a `FromStr` implementation in the common/install
logic), there is no longer any magically derived implicit default value
for a `PackageTarget`.

This work surfaced a few takeaways:

* For a lot of installation-related operations, a `PackageIdent` and
`PackageTarget` are tightly coupled and should appear together. This
reinforces a previous suspicion that we're missing new a type which
represents these 2 concepts in one place. This refactoring work has an
eye towards that future work by always placing `ident` and `target`
together in function calls.
* Most correct values for the package target end up being
`PackageTarget::active_target()` in the various call sites surfaced in
this work. This makes sense, in that the Supervisor must only ever deal
with packages matching its compiled in `PackageTarget` value. In a
similar fashion, a lot of CLI operations follow the same reasoning when
it installs supporting software such as the Studio, package exporters,
the Launcher, and the Supervisor.
* Moving implied defaults out to the edges allows the inner business
logic to be less `Option` obsessed and timid (i.e. constantly checking
whether or not a value has been provided).

Closes #5516
References habitat-sh/builder#612

Signed-off-by: Fletcher Nichol <fnichol@nichol.ca>