Add an error for an invalid field in the protocol message
I'm going to validate some of the fields and I didn't want to use
ProtocolMismatch because that error is about a missing field.

An example of validation would be if the port number in the message is
in valid range, because the field is a 32-bit integer (signed or
unsigned) and valid port number is in range of 16 bit unsigned
integer.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>