Discover host address of a member immediately
Do this now instead of waiting for the first SWIM message to figure
out the IP. We want this because the host address will be used to
figure out if the SWIM message came from other private network.

Also, supervisor was already using this upfront address discovery for
its service rumors, so we are just being consistent here.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>