Add zones
Zone is butterfly server's view of the private network. So servers
within a private network will get the same zone. And the zones will be
different across private networks.

Multiple servers within a single private network may start with
different zones, but over time they will converge on a single
zone. The zone with the highest ID wins (ID being a UUID). The zones
contain information about other zones of the same private network in
the successor and predecessors fields.

In future, the maintainer of a superseded zone should mark it as dead
after some period, so the list of superseded zones could shrink. This
is not yet implemented.

Zones have parent-children relationships between themselves. Servers
in child zone can initiate communication directly with a server in
parent zone, but not the other way around. That maps to the same
situation we have in NATs - servers behind a NAT can initiate
communication to servers outside NAT, but not the other way around.

Zones have a maintainer, which is a server that generated the zone
first. Currently the zone maintainership is not very useful, but might
be later if we will need to implement e.g. handling the case when the
parent zone is not really a parent zone, but a grandparent, so it
would need to be dropped from the parent info in favor of some other
zone that sits between our zone and the old parent, now-grandparent,
zone.

The zone negotiation process is about figuring out if the server is in
the same zone as the server that just sent a message to it. To be able
to do it, I added a "To" member to "Ping" and "Ack" messages that the
receiver of a message can use to figure out if the message was sent to
the same address as it's own deduced one. With that information,
together with the information about the server from the "From" field,
the server figure out whether to join the sender's zone or to keep
using its own (or generate a new one, when it is being talked to for
the first time).

Member struct contains information about the additional addresses it
is available under some zone. That information needs to be provided by
the user of the supervisor. The address can be discovered, but I'm not
really sold on the idea - it may interfere with the zone negotiation,
since they yet-undiscovered address can't be used for figuring out
whether the message came from other private network or from the same
one.

The additional addresses information also contain the "tag" field. It
will be later used by the supervisor in generating a census. For now
the only important information is that these tags need to be unique
only in the scope of a single supervisor.

Signed-off-by: Krzesimir Nowak <krzesimir@kinvolk.io>